move attedance check page view to its own widget
